# api/api-docs/build/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Xpense web api
  description: Xpense web api.
servers:
  - url: 'http://localhost:7000/api'
  - url: 'https://dev.api.xpense.bulldozerinc.com/api'
paths:
...
  '/org/{orgId}/sheets':
    get:
      summary: Get user organization sheets
      tags:
        - Organization
      operationId: getSheets
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Current user organization sheets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sheet'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: create sheet
      tags:
        - Sheets
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
      operationId: createSheet
      requestBody:
        description: create sheet body request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sheet'
      responses:
        '201':
          description: Sheet created
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
  '/org/{orgId}/sheets/{sheetId}':
    get:
      summary: get sheet by id
      tags:
        - Organization
        - Sheet
      operationId: getSheetById
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: sheetId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Get sheet by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sheet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: delete sheet
      tags:
        - Organization
        - Sheet
      operationId: deleteSheet
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: sheetId
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Sheet successfully delete.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /users:
    post:
      summary: Create new user
      tags:
        - Users
      operationId: createUser
      requestBody:
        description: Register user DTO object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      summary: List of users by organization
      tags:
        - Users
      operationId: usersBYorganization
      parameters:
        - in: query
          name: orgId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: List of filtered users by organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  '/users/{userId}':
    patch:
      summary: Change user role
      tags:
        - Users
      operationId: changeUserRole
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
      requestBody:
        description: Role Name
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roleName:
                  type: string
      responses:
        '204':
          description: Role successfully associated with user.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /roles:
    get:
      summary: List of roles
      tags:
        - Users
      operationId: getRoles
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Roles'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  '/org/{orgId}/taxes':
    get:
      summary: Get list of taxes
      tags:
        - Taxes
        - Organization
      operationId: getTaxes
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: List of taxes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Taxes'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      summary: Create new tax
      tags:
        - Organization
        - Tax
      operationId: createTax
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
      requestBody:
        description: Tax DTO object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTax'
      responses:
        '201':
          description: Tax created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tax'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
  '/org/{orgId}/taxes/{taxId}':
    get:
      summary: Get tax by id
      tags:
        - Taxes
        - Organization
      operationId: getTax
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: taxId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Tax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tax'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      summary: Update tax
      tags:
        - Organization
        - Tax
      operationId: updateTax
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: taxId
          schema:
            type: string
          required: true
      requestBody:
        description: Tax DTO object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTax'
      responses:
        '204':
          description: Tax successfully updated.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete tax
      tags:
        - Organization
        - Tax
      operationId: deleteTax
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: taxId
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Tax successfully delete.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
  '/sheets/{sheetId}/attachment':
    post:
      summary: upload attachment to sheet
      tags:
        - Sheets
      parameters:
        - in: path
          name: sheetId
          schema:
            type: string
          required: true
      operationId: insertAttachment
      requestBody:
        description: Upload attatchment file
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Upload'
      responses:
        '200':
          description: uploaded
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  '/sheets/{sheetId}/expenses/{expenseId}/attachment':
    post:
      summary: upload attachment to expense
      tags:
        - Expenses
      parameters:
        - in: path
          name: sheetId
          schema:
            type: string
          required: true
        - in: path
          name: expenseId
          schema:
            type: string
          required: true
      operationId: insertAttachment
      requestBody:
        description: Upload attatchment file
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Upload'
      responses:
        '200':
          description: uploaded
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Profile:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
        displayName:
          type: string
        lastName:
          type: string
        firstName:
          type: string
        photoURL:
          type: string
        defaultCurrency:
          type: string
        phoneNumber:
          type: string
        organizations:
          type: array
          items:
            type: string
        roleName:
          type: string
    CreateUser:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        displayName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        photoURL:
          type: string
        phoneNumber:
          type: string
...
    Sheet:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        status:
          type: string
        creationDate:
          type: string
          format: date-time
        submittedOn:
          type: string
          format: date-time
        exported:
          type: boolean
        total:
          type: integer
...
    Users:
      type: array
      items:
        $ref: '#/components/schemas/Profile'
    Roles:
      type: array
      items:
        type: object
        properties:
          id:
            type: string
          name:
            type: string
          description:
            type: string
            nullable: true
...
    Taxes:
      type: array
      items:
        $ref: '#/components/schemas/Tax'
    Tax:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        defaultRate:
          type: integer
          example: 0.09975
        i18n:
          type: string
    CreateTax:
      type: object
      properties:
        code:
          type: string
        defaultRate:
          type: integer
          example: 0.09975
        i18n:
          type: string
    Upload:
      type: object
      properties:
        file:
          type: string
          format: binary
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
    NotFound:
      description: The specified resource was not found
    BadRequest:
      description: Bad request
security:
  - bearerAuth: []